+++ src/main/java/com/ngo/fundraiser/entity/Donor.java
@@ -10,7 +10,7 @@ public class Donor {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long donorId;
+    private int donorId;
 
     private String name;
     private String address;
@@ -22,15 +22,15 @@ public class Donor {
     public Donor() {
     }
 
-    public Donor(Long donorId) {
+    public Donor(int donorId) {
         this.donorId = donorId;
     }
 
-    public Long getDonorId() {
+    public int getDonorId() {
         return donorId;
     }
 
-    public void setDonorId(Long donorId) {
+    public void setDonorId(int donorId) {
         this.donorId = donorId;
     }
